commit
changed color green of selected scored questins, fixed accordian arrow, created two parts to the questions which broke up the accordians in to 1 and 2, This caused a need for an update on the slectors for the question class name and color indicator. User and question routes now updated and fixed. user topics still hard coded in app.js for now